move quickcache to its own package
(still has a small corehq dependency, to be removed
