+++ packages/quickcache/__init__.py
@@ -4,13 +4,9 @@ from .quickcache import (
 )
 from .django_quickcache import quickcache
 
-from .deprecated import SkippableQuickCache, skippable_quickcache
-
 
 __all__ = [
     'generic_quickcache',
-    'quickcache',
-    'skippable_quickcache',
+    'quickcache'
     'QuickCacheHelper',
-    'SkippableQuickCache',
 ]
